requirements_static.txt
Flask>=2.2
Frozen-Flask>=0.18
python-dotenv>=1.0
Flask-SQLAlchemy>=3.0
